Checking if executables exist before assuming.

+++ osx/defaults.zsh
@@ -1,2 +1,4 @@
 # Allow text selection in the Quick Look window
-defaults write com.apple.finder QLEnableTextSelection -bool true && killall Finder
+if hash defaults 2>/dev/null; then
+  defaults write com.apple.finder QLEnableTextSelection -bool true && killall Finder
+fi
